requirements.txt
python-daemon
PyUserInput
python3-xlib
